+++ changelog.md
@@ -5,6 +5,7 @@
 * [Bug] - Fix NightBloom File ID extraction regex following API change
 * [Exp] - Add experimental support for multi-sourceset projects
 * [Feat] - Add support for new snapshot version formats for CurseForge
+* [Feat] - Add option to not set GitHub release as latest
 
 ### Version 2.1.7